% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ort_codesign}
\alias{ort_codesign}
\title{ort_codesign}
\usage{
ort_codesign(lib_dir)
}
\arguments{
\item{lib_dir}{where the libraries are}
}
\value{
invisible lib path
}
\description{
for macs, signs the downloaded binaries
}
\examples{
\dontrun{ort_codesign(file.path(ort_install_dir(), "lib"))}
}
