% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ort_download}
\alias{ort_download}
\title{ort_download}
\usage{
ort_download(url, dest_dir)
}
\arguments{
\item{url}{where to download from}

\item{dest_dir}{where to download to}
}
\value{
invisible dest_dir
}
\description{
ort_download
}
\examples{
\donttest{
if (!ort_is_installed()){
  ort_download(ort_binary_url(), ort_install_dir())}
}
}
