% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{ort_infer_raw}
\alias{ort_infer_raw}
\title{ort_infer_raw}
\usage{
ort_infer_raw(session, input)
}
\arguments{
\item{session}{ORT session}

\item{input}{tensor to infer on}
}
\value{
ORT output
}
\description{
ort_infer_raw
}
\examples{
\dontrun{ort_infer_raw(session, input)}
}
