% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{ort_is_installed}
\alias{ort_is_installed}
\title{ort_is_installed}
\usage{
ort_is_installed()
}
\value{
boolean for if you have onnx runtime or not
}
\description{
ort_is_installed
}
\examples{
ort_is_installed()
}
