% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{ort_session}
\alias{ort_session}
\title{session.R}
\usage{
ort_session(
  path,
  type = "detection",
  provider = "cpu",
  cache_dir = NULL,
  threads = 0L,
  opt_level = 99L
)
}
\arguments{
\item{path}{where the .onnx is located}

\item{type}{detection/classificaton/segmentation}
}
\value{
ORT session object
}
\description{
session.R
}
\examples{
ort_session('./yolov8n.onnx', 'detection')
}
