% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{print.ort_session}
\alias{print.ort_session}
\title{print.ort_session}
\usage{
\method{print}{ort_session}(x, ...)
}
\arguments{
\item{x}{session object}

\item{...}{extra params}
}
\value{
invisible session
}
\description{
print.ort_session
}
\examples{
\donttest{
  if (file.exists("model.onnx")) {
    session <- ort_session("model.onnx")
    print(session)
  }
}
}
