% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{ort_session}
\alias{ort_session}
\title{session.R}
\usage{
ort_session(
  path,
  type = "detection",
  provider = "cpu",
  cache_dir = NULL,
  threads = 0L,
  opt_level = 99L
)
}
\arguments{
\item{path}{where the .onnx is located}

\item{type}{detection/classificaton/segmentation}

\item{provider}{cpu or coreml (only for apple silicon)}

\item{cache_dir}{where coreml cache be directed}

\item{threads}{=  0 for all, integer otherwise}

\item{opt_level}{= 99 for all ops, 1 for basics}
}
\value{
ORT session object
}
\description{
session.R
}
\examples{
\dontrun{ort_session('./yolov8n.onnx', 'detection')}
}
